frontend/requirements.txt
# General
python-dotenv
setuptools
toml

# Testing
pytest
pytest-cov
pytest-asyncio

# FastAPI
fastapi
uvicorn
pydantic

# Streamlit
streamlit
streamlit-folium

# HTTP
requests
aiohttp

# Database
sqlalchemy
psycopg2-binary
supabase

# Data manipulation
numpy==2.0.2
pandas

# Machine Learning
tensorflow==2.18.0
keras==3.8.0
scikit-learn
joblib
prophet

# Data Visualization
folium
matplotlib
plotly